+++ crates/bindings-typescript/src/server/type_builders.ts
@@ -13,7 +13,7 @@ import {
   type TimestampAlgebraicType,
 } from '..';
 import type { OptionAlgebraicType } from '../lib/option';
-import { addType, MODULE_DEF } from './runtime';
+import { addType, MODULE_DEF } from './schema';
 import type { CoerceRow } from './table';
 import { set, type Set } from './type_util';
